Organized robot\config
