--- src/main/java/frc/robot/config/RobotConfigStub.java
@@ -1,3 +1,0 @@
-package frc.robot.config;
-
-public class RobotConfigStub extends RobotConfig {}
