+++ src/main/java/frc/robot/config/game/reefscape2025/RobotConfigPhoenix.java
@@ -1,8 +1,9 @@
-package frc.robot.config;
+package frc.robot.config.game.reefscape2025;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 import edu.wpi.first.math.util.Units;
 import frc.robot.commands.game.crescendo2024.auto.AutoNamedCommands;
+import frc.robot.config.RobotConfig;
 import frc.robot.subsystems.arm.ArmSubsystem;
 import frc.robot.subsystems.drive.Drive;
 import frc.robot.subsystems.drive.DriveBase;
